+++ lib/src/v2/adyen_session_component.dart
@@ -1,4 +1,6 @@
 import 'package:adyen_checkout/src/adyen_checkout.dart';
+import 'package:adyen_checkout/src/common/model/before_submit.dart';
+import 'package:adyen_checkout/src/common/model/checkout.dart';
 import 'package:adyen_checkout/src/common/model/payment_result.dart';
 import 'package:adyen_checkout/src/common/model/result_code.dart';
 import 'package:adyen_checkout/src/generated/platform_api.g.dart';
@@ -6,8 +8,9 @@ import 'package:adyen_checkout/src/util/constants.dart';
 import 'package:adyen_checkout/src/util/dto_mapper.dart';
 import 'package:adyen_checkout/src/v2/adyen_base_component.dart';
 
-class AdyenSessionComponent extends AdyenBaseComponent {
-  final SessionDTO session;
+class AdyenSessionComponent extends AdyenBaseComponent
+    implements SessionCheckoutFlutterInterface {
+  final SessionCheckout sessionCheckout;
 
   @override
   final String componentId = "SESSION_ADYEN_COMPONENT";
@@ -18,7 +21,7 @@ class AdyenSessionComponent extends AdyenBaseComponent {
   AdyenSessionComponent({
     super.key,
     required super.checkoutConfiguration,
-    required this.session,
+    required this.sessionCheckout,
     required super.paymentMethod,
     required super.paymentMethodTxVariant,
     required super.onPaymentResult,
@@ -28,11 +31,13 @@ class AdyenSessionComponent extends AdyenBaseComponent {
     super.adyenLogger,
     super.onBinLookup,
     super.onBinValue,
-  });
+  }) {
+    SessionCheckoutFlutterInterface.setUp(this);
+  }
 
   @override
   Map<String, dynamic> get creationParams => <String, dynamic>{
-        Constants.sessionKey: session,
+        Constants.sessionKey: sessionCheckout.toDTO(),
         Constants.checkoutConfigurationKey: checkoutConfiguration,
         Constants.paymentMethodKey: paymentMethod,
         Constants.paymentMethodTxVariantKey: paymentMethodTxVariant,
@@ -51,6 +56,17 @@ class AdyenSessionComponent extends AdyenBaseComponent {
       sessionResult: paymentResultDTO?.result?.sessionResult ?? "",
       resultCode: resultCode,
     ));
+  }
+
+  @override
+  Future<BeforeSubmitResultDTO> onBeforeSubmit(BeforeSubmitDataDTO data) async {
+    final onBeforeSubmit = sessionCheckout.onBeforeSubmit;
+    if (onBeforeSubmit == null) {
+      return BeforeSubmitResultDTO(isAborted: false, data: data);
+    }
+
+    final BeforeSubmitResult result = await onBeforeSubmit(data.fromDTO());
+    return result.toDTO();
   }
 
   void _resetSession() => AdyenCheckout.session.clear();